pyproject.toml
[tool.poetry]
name = "fastapi-neo4j"
version = "0.1.0"
description = ""
authors = ["simon.islam <simon.islam@brilliant.com.bd>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.78.0"
uvicorn = "^0.17.6"
gunicorn = "^20.1.0"
neo4j = "^4.4.3"
passlib = "^1.7.4"
PyJWT = "^2.4.0"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
pytest-cov = "^3.0.0"
black = "^22.3.0"
isort = "^5.10.1"
flake8 = "^4.0.1"
bandit = "^1.7.4"
safety = "^1.10.3"
pre-commit = "^2.19.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
